+++ packages/components/modules/payments/web/AvailableSubscriptions/index.tsx
@@ -1,14 +1,14 @@
 import { useState } from 'react'
 
 import { LoadingState } from '@baseapp-frontend/design-system/components/web/displays'
-import { CheckMarkIcon } from '@baseapp-frontend/design-system/components/web/icons'
 import { useResponsive } from '@baseapp-frontend/design-system/hooks/web'
 
-import { Box, Button, Chip, ToggleButton, ToggleButtonGroup, Typography } from '@mui/material'
-import Image from 'next/image'
+import { Box, ToggleButton, ToggleButtonGroup } from '@mui/material'
 import { useRouter } from 'next/router'
 
 import useStripeHook from '../hooks/useStripeHook'
+import { IProduct } from '../types'
+import SubscriptionCard from './SubscriptionCard'
 
 const AvailableSubscriptions = () => {
   const [selectedTerm, setSelectedTerm] = useState<'monthly' | 'yearly'>('monthly')
@@ -48,67 +48,16 @@ const AvailableSubscriptions = () => {
         </ToggleButtonGroup>
       </Box>
       <Box display="flex" gap={2} width="100%" height="100%" flexWrap="wrap">
-        {subs?.map((sub) => {
-          const dolarPrice = Math.floor((sub.defaultPrice?.unitAmount ?? 0) / 100)
-          const centsPrice = ((sub.defaultPrice?.unitAmount ?? 0) % 100).toString().padStart(2, '0')
-          return (
-            <Box
-              key={sub.id}
-              sx={{
-                display: 'flex',
-                flexDirection: 'column',
-                border: '1px solid',
-                borderColor: 'divider',
-                borderRadius: 2,
-                padding: 2,
-                gap: 2,
-                width: '100%',
-                minWidth: '300px',
-                maxWidth: smDown ? '100%' : '300px',
-                flex: '1 1 300px',
-              }}
-            >
-              {sub?.images?.length > 0 && (
-                <Image src={sub.images[0] ?? ''} alt={sub.name} width={28} height={28} />
-              )}
-              <Box display="flex" gap={1} alignItems="flex-end">
-                <Typography variant="h4">{sub.name}</Typography>
-                {isActive && <Chip label="Active" color="success" variant="soft" />}
-              </Box>
-              <Box display="flex" gap={1} alignItems="flex-end">
-                <Typography variant="h4" color="text.secondary">
-                  $
-                </Typography>
-                <Typography variant="h2">{dolarPrice}</Typography>
-                <Typography variant="body1">.{centsPrice}</Typography>
-                <Typography variant="body1" color="text.secondary">
-                  USD/{selectedTerm}
-                </Typography>
-              </Box>
-              <Typography variant="body2">{sub.description}</Typography>
-              {isActive ? (
-                <Button
-                  variant="soft"
-                  color="inherit"
-                  onClick={() => router.push('/user/settings?tab=subscription')}
-                >
-                  Manage Subscription
-                </Button>
-              ) : (
-                <Button variant="contained" color="inherit">
-                  Subscribe
-                </Button>
-              )}
-              <Box display="flex" flexDirection="column" gap={2}>
-                {sub.marketingFeatures?.map((feature) => (
-                  <Typography variant="body2" color="text.secondary" key={feature.name}>
-                    <CheckMarkIcon color="inherit" /> {feature.name}
-                  </Typography>
-                ))}
-              </Box>
-            </Box>
-          )
-        })}
+        {subs?.map((sub) => (
+          <SubscriptionCard
+            key={sub.id}
+            sub={sub as IProduct}
+            isActive={isActive}
+            smDown={smDown}
+            selectedTerm={selectedTerm}
+            onClick={() => router.push('/user/settings?tab=subscription')}
+          />
+        ))}
       </Box>
     </>
   )
